Domain Object was changed: The DeleteDomainObject method is now aware if the lockMode is on or off. If it is off then nothing is done, since the intention of this method is to delete the master association of this object with the persistence. Practical use: it now works calling the deleteDomainObject inside a test.

+++ src/net/sourceforge/fenixedu/domain/DomainObject.java
@@ -77,11 +77,13 @@ public abstract class DomainObject extends DomainObject_Base {
     }
 
     protected final void deleteDomainObject() {
-        try {
-            PersistenceSupportFactory.getDefaultPersistenceSupport()
-                    .getIPersistentObject().deleteByOID(this.getClass(), getIdInternal());
-        } catch (ExcepcaoPersistencia e) {
-            throw new Error("Couldn't delete object", e);
+        if (lockMode) {
+            try {
+                PersistenceSupportFactory.getDefaultPersistenceSupport()
+                        .getIPersistentObject().deleteByOID(this.getClass(), getIdInternal());
+            } catch (ExcepcaoPersistencia e) {
+                throw new Error("Couldn't delete object", e);
+            }
         }
     }
